revert @coveo/create-atomic-commons
https://coveord.atlassian.net/browse/CDX-1400
