+++ packages/cli-e2e/setup/utils/verdaccio.ts
@@ -12,7 +12,6 @@ const verdaccioRegistry = 'http://localhost:4873';
 
 const verdaccioedPackages = [
   '@coveo/atomic-component-health-check',
-  '@coveo/create-atomic-commons',
   '@coveo/create-atomic-component',
   '@coveo/create-atomic-component-project',
   '@coveo/create-atomic-result-component',
