+++ packages/ui/atomic/create-atomic-result-component/package.json
@@ -22,7 +22,6 @@
     "access": "public"
   },
   "dependencies": {
-    "@coveo/create-atomic-commons": "0.0.1",
     "@coveo/create-atomic-component-project": "0.2.1"
   },
   "devDependencies": {
