+++ packages/ui/atomic/create-atomic-result-component/tests/test.spec.ts
@@ -9,6 +9,7 @@ import treeKill from 'tree-kill-promise';
 import {SpawnSyncReturns} from 'child_process';
 
 const PACKAGE_NAME = '@coveo/create-atomic-result-component';
+// const UTILS_PACKAGE_NAME = '@coveo/create-atomic-commons';
 
 describe(PACKAGE_NAME, () => {
   let verdaccioProcess: ChildProcess;
@@ -18,7 +19,10 @@ describe(PACKAGE_NAME, () => {
   let verdaccioUrl: string;
 
   beforeAll(async () => {
-    ({verdaccioUrl, verdaccioProcess} = await startVerdaccio(PACKAGE_NAME));
+    ({verdaccioUrl, verdaccioProcess} = await startVerdaccio([
+      PACKAGE_NAME,
+      // UTILS_PACKAGE_NAME, // TODO: include @coveo/create-atomic-commons
+    ]));
     tempDirectory = dirSync({unsafeCleanup: true, keep: true});
     npmConfigCache = join(tempDirectory.name, 'npm-cache');
     mkdirSync(npmConfigCache);
@@ -94,26 +98,27 @@ describe(PACKAGE_NAME, () => {
       it('should ouptut a confirmation message upon success', () => {
         expect(commandOutput.stdout.toString()).toMatchSnapshot();
       });
+    });
 
-      it('should be able to install all deps without issues', () => {
-        const output: any = npmSync(['install'], {
+    it('should be able to install all deps without issues', () => {
+      expect(
+        npmSync(['install'], {
           cwd: testDirectory,
           env: {
             ...process.env,
             npm_config_registry: verdaccioUrl,
             npm_config_cache: npmConfigCache,
           },
-        }).status;
-        expect(output).toBe(0);
-      });
+        }).status
+      ).toBe(0);
+    });
 
-      it('should be able to build without issues', () => {
-        expect(
-          npmSync(['run', 'build', '-w', packageName], {
-            cwd: testDirectory,
-          }).status
-        ).toBe(0);
-      });
+    it('should be able to build without issues', () => {
+      expect(
+        npmSync(['run', 'build', '-w', packageName], {
+          cwd: testDirectory,
+        }).status
+      ).toBe(0);
     });
   });
 });
