--- utils/create-atomic-commons/.eslintrc.json
@@ -1,4 +1,0 @@
-{
-  "extends": "../../base.eslintrc",
-  "parserOptions": {"project": "./tsconfig.json"}
-}
