--- utils/create-atomic-commons/.gitignore
@@ -1,3 +1,0 @@
-coverage
-node_modules
-dist
