--- utils/create-atomic-commons/jest.config.js
@@ -1,14 +1,0 @@
-export default {
-  preset: 'ts-jest/presets/default-esm',
-  extensionsToTreatAsEsm: ['.ts'],
-  moduleNameMapper: {
-    '^(\\.{1,2}/.*)\\.js$': '$1',
-  },
-  testEnvironment: 'node',
-  verbose: true,
-  collectCoverage: true,
-  clearMocks: true,
-  silent: true,
-  testTimeout: 60e3,
-  testPathIgnorePatterns: ['/node_modules/', '/dist/'],
-};
