--- utils/create-atomic-commons/package.json
@@ -1,44 +1,0 @@
-{
-  "name": "@coveo/create-atomic-commons",
-  "type": "module",
-  "version": "0.0.1",
-  "description": "Common utils for atomic:component command",
-  "main": "./dist/index.js",
-  "author": "Coveo",
-  "homepage": "https://github.com/coveo/cli#readme",
-  "publishConfig": {
-    "access": "public"
-  },
-  "license": "Apache-2.0",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.js"
-    }
-  },
-  "scripts": {
-    "release:phase1": "npx -p=@coveord/release npm-publish",
-    "build": "rimraf dist && tsc -b tsconfig.build.json",
-    "prepublishOnly": "npm run build",
-    "pretest": "rimraf dist && tsc -b tsconfig.json",
-    "test": "jest"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/coveo/cli.git"
-  },
-  "bugs": {
-    "url": "https://github.com/coveo/cli/issues"
-  },
-  "files": [
-    "dist"
-  ],
-  "devDependencies": {
-    "@types/jest": "^29.5.0",
-    "@types/node": "^18.14.6",
-    "jest": "^29.5.0",
-    "rimraf": "^5.0.0",
-    "ts-jest": "^29.0.5",
-    "typescript": "^4.9.5"
-  }
-}
