--- utils/create-atomic-commons/src/index.ts
@@ -1,3 +1,0 @@
-export * from './console.js';
-export * from './transformers.js';
-export * from './validation.js';
