--- utils/create-atomic-commons/tsconfig.build.json
@@ -1,8 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "sourceMap": false,
-    "declaration": false
-  },
-  "exclude": ["node_modules", "src/**/*.spec.ts"]
-}
