--- utils/create-atomic-commons/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "skipLibCheck": true,
-    "outDir": "dist",
-    "module": "Node16",
-    "target": "ESNext",
-    "moduleResolution": "Node16",
-    "sourceMap": true,
-    "lib": ["ESNext"],
-    "declaration": true,
-    "esModuleInterop": true
-  },
-  "exclude": ["node_modules"]
-}
